0.0.5 - Changes and Updated to 1.6
0.0.5 - Changes and Updated to 1.6

It has been deemed that this mod needed no changes to be compatible with 1.6.

I organized some of the scripts better.

I moved ClassCopy to ClassUtility.

I added GetField and GetRecursiveField to ClassUtility.

I moved all files from 1.5 folder to just the main folder since currently there is no need to be so specific.