0.0.8 - Minor Fixes
Fixed a minor issue in Mastery_StatPart that might have caused a error in a few situations.

Although no one has run into a error because of it and to my knowledge never would have, but it is better to be safe than sorry.

Also a minor change to ActionEvent on Level_Comp to allow the Action_Manager to actually work without causing problems.