0.0.9 - Minor Improvements
Made it to where you no longer need to create the value(aka override a mastery) to enable/disable a mastery for better storage usage.

Renamed IsIgnored to IsEnabled in configs(for defs/extensions it's isEnabledByDefault).

Did A Few Minor Changes To the Settings Backend.

UIUtility Now has the option for a foldout with a checkbox.

PLEASE NOTE: Any mods that previously used this will need to be updated to work with this latest version.